refactor: decompose App and TheoryControls components
- Decomposed TheoryControls into ScaleSelector, ChordOverlayControls, and KeyExplorer.
- Extracted HelpModal, SummaryRibbon, and MainLayoutWrapper from App.tsx.
- Introduced hiddenNotesAtom for centralized note visibility state.
- Updated components to consume domain-specific hooks directly, reducing prop drilling.
- Cleaned up ExpandedControlsPanel and updated integration tests.
