+++ src/components/MainLayoutWrapper.tsx
@@ -1,0 +1,69 @@
+import { type ReactNode } from "react";
+
+interface MainLayoutWrapperProps {
+  children: ReactNode;
+  header: ReactNode;
+  summary?: ReactNode;
+  controlsPanel?: ReactNode;
+  mobileTabs?: ReactNode;
+  helpModal?: ReactNode;
+  settingsOverlay?: ReactNode;
+  versionBadge?: ReactNode;
+  layoutTier: string;
+  layoutVariant: string;
+  isChordActive: boolean;
+  showHeaderSubtitle: boolean;
+  compactHeaderActions: boolean;
+  fullWidthOverlay: boolean;
+  showSummary: boolean;
+  showControlsPanel: boolean;
+  showMobileTabs: boolean;
+}
+
+export function MainLayoutWrapper({
+  children,
+  header,
+  summary,
+  controlsPanel,
+  mobileTabs,
+  helpModal,
+  settingsOverlay,
+  versionBadge,
+  layoutTier,
+  layoutVariant,
+  isChordActive,
+  showHeaderSubtitle,
+  compactHeaderActions,
+  fullWidthOverlay,
+  showSummary,
+  showControlsPanel,
+  showMobileTabs,
+}: MainLayoutWrapperProps) {
+  return (
+    <div
+      className="app-container"
+      data-layout-tier={layoutTier}
+      data-layout-variant={layoutVariant}
+      data-chord-active={isChordActive ? "true" : undefined}
+      data-header-subtitle={showHeaderSubtitle ? "visible" : "hidden"}
+      data-header-actions={compactHeaderActions ? "compact" : "default"}
+      data-full-width-overlay={fullWidthOverlay ? "true" : "false"}
+    >
+      {header}
+
+      {showSummary && <div className="summary-shell">{summary}</div>}
+
+      {helpModal}
+
+      <main className="main-fretboard">{children}</main>
+
+      {showControlsPanel && controlsPanel}
+
+      {showMobileTabs && mobileTabs}
+
+      {versionBadge}
+
+      {settingsOverlay}
+    </div>
+  );
+}
