+++ src/components/TheoryControls.test.tsx
@@ -1,39 +1,31 @@
 // @vitest-environment jsdom
-import { describe, it, expect, vi } from "vitest";
+import { describe, it, expect, beforeEach } from "vitest";
 import { fireEvent, render, screen } from "@testing-library/react";
-import type { ComponentProps } from "react";
+import { createStore, Provider } from "jotai";
 import { TheoryControls } from "./TheoryControls";
+import {
+  rootNoteAtom,
+  scaleNameAtom,
+  scaleBrowseModeAtom,
+  chordTypeAtom,
+  viewModeAtom,
+  focusPresetAtom,
+} from "../store/atoms";
 
-function makeProps(): ComponentProps<typeof TheoryControls> {
+function renderWithStore(ui: React.ReactElement, store = createStore()) {
   return {
-    rootNote: "C",
-    setRootNote: vi.fn(),
-    scaleName: "Major",
-    setScaleName: vi.fn(),
-    scaleBrowseMode: "parallel",
-    setScaleBrowseMode: vi.fn(),
-    chordType: null,
-    setChordType: vi.fn(),
-    chordRoot: "C",
-    setChordRoot: vi.fn(),
-    linkChordRoot: true,
-    setLinkChordRoot: vi.fn(),
-    viewMode: "compare",
-    setViewMode: vi.fn(),
-    focusPreset: "all",
-    setFocusPreset: vi.fn(),
-    customMembers: [],
-    setCustomMembers: vi.fn(),
-    availableFocusPresets: ["all", "rootless", "custom"],
-    chordMembers: [],
-    hasOutsideChordMembers: false,
-    useFlats: false,
+    ...render(<Provider store={store}>{ui}</Provider>),
+    store,
   };
 }
 
 describe("TheoryControls", () => {
+  beforeEach(() => {
+    localStorage.clear();
+  });
+
   it("renders root, family, and browse controls with chord overlay collapsed", () => {
-    render(<TheoryControls {...makeProps()} />);
+    renderWithStore(<TheoryControls />);
 
     expect(screen.getByText("Root")).toBeInTheDocument();
     expect(
@@ -49,72 +41,67 @@ describe("TheoryControls", () => {
     expect(
       screen.getByRole("button", { name: /Next Mode/i }),
     ).toBeInTheDocument();
+    expect(screen.getByRole("combobox", { name: "Mode" })).toBeInTheDocument();
     expect(
-      screen.getByRole("combobox", { name: "Mode" }),
+      screen.getByRole("button", { name: /Chord Overlay/i }),
     ).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: /Chord Overlay/i })).toBeInTheDocument();
     expect(screen.queryByText("Chord Type")).not.toBeInTheDocument();
   });
 
   it("switches families using the scale family select", () => {
-    const props = makeProps();
-    render(<TheoryControls {...props} />);
+    const { store } = renderWithStore(<TheoryControls />);
 
     fireEvent.change(screen.getByRole("combobox", { name: "Scale Family" }), {
       target: { value: "Pentatonic" },
     });
 
-    expect(props.setScaleName).toHaveBeenCalledWith("Minor Pentatonic");
+    expect(store.get(scaleNameAtom)).toBe("Minor Pentatonic");
   });
 
   it("switches the active mode using the browse select in parallel mode", () => {
-    const props = makeProps();
-    render(<TheoryControls {...props} />);
+    const { store } = renderWithStore(<TheoryControls />);
 
     fireEvent.change(screen.getByRole("combobox", { name: "Mode" }), {
       target: { value: "C Dorian" },
     });
 
-    expect(props.setScaleName).toHaveBeenCalledWith("Dorian");
-    expect(props.setRootNote).toHaveBeenCalledWith("C");
+    expect(store.get(scaleNameAtom)).toBe("Dorian");
+    expect(store.get(rootNoteAtom)).toBe("C");
   });
 
   it("switches root and mode together when relative browsing is active", () => {
-    const props = makeProps();
-    render(
-      <TheoryControls
-        {...props}
-        rootNote="C"
-        scaleName="Major"
-        scaleBrowseMode="relative"
-      />,
-    );
+    const store = createStore();
+    store.set(rootNoteAtom, "C");
+    store.set(scaleNameAtom, "Major");
+    store.set(scaleBrowseModeAtom, "relative");
+
+    renderWithStore(<TheoryControls />, store);
 
     fireEvent.change(screen.getByRole("combobox", { name: "Mode" }), {
       target: { value: "D Dorian (2nd Mode)" },
     });
 
-    expect(props.setRootNote).toHaveBeenCalledWith("D");
-    expect(props.setScaleName).toHaveBeenCalledWith("Dorian");
+    expect(store.get(rootNoteAtom)).toBe("D");
+    expect(store.get(scaleNameAtom)).toBe("Dorian");
   });
 
   it("hides the browse-mode toggle for variant families", () => {
-    render(
-      <TheoryControls
-        {...makeProps()}
-        scaleName="Minor Pentatonic"
-        scaleBrowseMode="relative"
-      />,
-    );
+    const store = createStore();
+    store.set(scaleNameAtom, "Minor Pentatonic");
+    store.set(scaleBrowseModeAtom, "relative");
 
-    expect(screen.queryByRole("button", { name: "Parallel" })).not.toBeInTheDocument();
+    renderWithStore(<TheoryControls />, store);
+
+    expect(
+      screen.queryByRole("button", { name: "Parallel" }),
+    ).not.toBeInTheDocument();
     expect(
       screen.getByRole("combobox", { name: "Variant" }),
     ).toBeInTheDocument();
   });
 
   it("expands the chord overlay controls on demand", () => {
-    render(<TheoryControls {...makeProps()} />);
+    renderWithStore(<TheoryControls />);
 
     fireEvent.click(screen.getByRole("button", { name: /Chord Overlay/i }));
 
@@ -122,9 +109,7 @@ describe("TheoryControls", () => {
   });
 
   it("shows the inline key explorer only after disclosure is opened", () => {
-    render(
-      <TheoryControls {...makeProps()} keyExplorer={<div>Key Wheel</div>} />,
-    );
+    renderWithStore(<TheoryControls keyExplorer={<div>Key Wheel</div>} />);
 
     expect(screen.queryByText("Key Wheel")).not.toBeInTheDocument();
     fireEvent.click(screen.getByRole("button", { name: /Circle of Fifths/i }));
@@ -132,74 +117,50 @@ describe("TheoryControls", () => {
   });
 
   it("shows View and Focus controls when a chord type is selected", () => {
-    const props = makeProps();
-    render(
-      <TheoryControls
-        {...props}
-        chordType="Major Triad"
-        availableFocusPresets={["all", "rootless", "custom"]}
-        chordMembers={[
-          { name: "root", semitone: 0, note: "C" },
-          { name: "3", semitone: 4, note: "E" },
-          { name: "5", semitone: 7, note: "G" },
-        ]}
-      />,
-    );
+    const store = createStore();
+    store.set(chordTypeAtom, "Major Triad");
+
+    renderWithStore(<TheoryControls />, store);
 
     expect(screen.getByText("View")).toBeInTheDocument();
     expect(screen.getByText("Focus")).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: "Scale + Chord" })).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: "Chord Only" })).toBeInTheDocument();
+    expect(
+      screen.getByRole("button", { name: "Scale + Chord" }),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole("button", { name: "Chord Only" }),
+    ).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "All" })).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "Rootless" })).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "Custom" })).toBeInTheDocument();
   });
 
   it("calls setViewMode when a view option is clicked", () => {
-    const props = makeProps();
-    render(
-      <TheoryControls
-        {...props}
-        chordType="Major Triad"
-        availableFocusPresets={["all", "rootless", "custom"]}
-        chordMembers={[]}
-      />,
-    );
+    const store = createStore();
+    store.set(chordTypeAtom, "Major Triad");
+
+    renderWithStore(<TheoryControls />, store);
 
     fireEvent.click(screen.getByRole("button", { name: "Chord Only" }));
-    expect(props.setViewMode).toHaveBeenCalledWith("chord");
+    expect(store.get(viewModeAtom)).toBe("chord");
   });
 
   it("calls setFocusPreset when a focus option is clicked", () => {
-    const props = makeProps();
-    render(
-      <TheoryControls
-        {...props}
-        chordType="Major Triad"
-        availableFocusPresets={["all", "rootless", "custom"]}
-        chordMembers={[]}
-      />,
-    );
+    const store = createStore();
+    store.set(chordTypeAtom, "Major Triad");
+
+    renderWithStore(<TheoryControls />, store);
 
     fireEvent.click(screen.getByRole("button", { name: "Rootless" }));
-    expect(props.setFocusPreset).toHaveBeenCalledWith("rootless");
+    expect(store.get(focusPresetAtom)).toBe("rootless");
   });
 
   it("shows custom member toggles when focusPreset is custom", () => {
-    const props = makeProps();
-    render(
-      <TheoryControls
-        {...props}
-        chordType="Major Triad"
-        focusPreset="custom"
-        availableFocusPresets={["all", "rootless", "custom"]}
-        chordMembers={[
-          { name: "root", semitone: 0, note: "C" },
-          { name: "3", semitone: 4, note: "E" },
-          { name: "5", semitone: 7, note: "G" },
-        ]}
-      />,
-    );
+    const store = createStore();
+    store.set(chordTypeAtom, "Major Triad");
+    store.set(focusPresetAtom, "custom");
+
+    renderWithStore(<TheoryControls />, store);
 
     expect(screen.getByText("Members")).toBeInTheDocument();
     expect(screen.getByRole("button", { name: "Root" })).toBeInTheDocument();
@@ -208,29 +169,28 @@ describe("TheoryControls", () => {
   });
 
   it("Outside view option is disabled when hasOutsideChordMembers is false", () => {
-    render(
-      <TheoryControls
-        {...makeProps()}
-        chordType="Major Triad"
-        availableFocusPresets={["all", "rootless", "custom"]}
-        chordMembers={[]}
-        hasOutsideChordMembers={false}
-      />,
-    );
+    const store = createStore();
+    store.set(rootNoteAtom, "C");
+    store.set(scaleNameAtom, "Major");
+    store.set(chordTypeAtom, "Major Triad");
+
+    renderWithStore(<TheoryControls />, store);
 
     expect(screen.getByRole("button", { name: "Outside" })).toBeDisabled();
   });
 
   it("Outside view option is enabled when hasOutsideChordMembers is true", () => {
-    render(
-      <TheoryControls
-        {...makeProps()}
-        chordType="Major Triad"
-        availableFocusPresets={["all", "rootless", "custom"]}
-        chordMembers={[]}
-        hasOutsideChordMembers={true}
-      />,
-    );
+    const store = createStore();
+    store.set(rootNoteAtom, "C");
+    store.set(scaleNameAtom, "Major");
+    store.set(chordTypeAtom, "Major 7th"); // This will have outside members if scale is major pentatonic? 
+    // No, Major 7th (C E G B) is in C Major (C D E F G A B).
+    // Let's use a chord that is definitely outside.
+    // Major 7th against Minor Pentatonic.
+    store.set(scaleNameAtom, "Minor Pentatonic"); // C Eb F G Bb
+    store.set(chordTypeAtom, "Major Triad"); // C E G -> E is outside.
+
+    renderWithStore(<TheoryControls />, store);
 
     expect(screen.getByRole("button", { name: "Outside" })).not.toBeDisabled();
   });
